Introduce ApolloServerPluginLandingPageDefault
As part of reducing Apollo Server's key dependencies on third-party
software (especially retired software) we are considering changing the
default landing page from GraphQL Playground to a simple splash page
welcoming the user to Apollo Server and providing them with links to
docs, to the Apollo Studio Sandbox, etc. (Sandbox is a forthcoming
feature that will allow you to use Explorer and other Studio features
without any need for a Studio account.)

This commit adds a new landing page plugin,
ApolloServerPluginLandingPageDefault, which serves this landing page
from a CDN. Note that despite its name, it is not actually made default
in this PR; we are likely to do that before 3.0.0 but after the Sandbox
feature linked from it is launched.